fixing merge errors

+++ bostonBusMap/src/boston/Bus/Map/transit/CommuterRailTransitSource.java
@@ -99,8 +99,8 @@ public class CommuterRailTransitSource implements TransitSource {
 
 	@Override
 	public void refreshData(RouteConfig routeConfig,
-			int selectedBusPredictions, int maxStops, float centerLatitude,
-			float centerLongitude,
+			int selectedBusPredictions, int maxStops, double centerLatitude,
+			double centerLongitude,
 			ConcurrentHashMap<Integer, BusLocation> busMapping,
 			String selectedRoute, RoutePool routePool, Directions directions,
 			Locations locationsObj) throws IOException,
